Object Documantation

Index

MultipathChannel.ChanPars

MultipathChannel.applyChan

MultipathChannel.initChan

Objects

MultipathChannel.applyChan — Method

MultipathChannel.applyChan(in, chan::ChanPars; initialtime)

Apply multipath channel based on chan.

Arguments

in: Input signal

chan::ChanPars: Channel parameters

initialtime: Time for generating Doppler shift

MultipathChannel.initChan — Method

MultipathChannel.initChan(maxdoppler, samplingrate, channelmodel; rmsdelay, maxorder, gainnormalization=true)

Generate a ChanPars type channel parameters.

Arguments

maxdoppler: Maximum Doppler frequency (Hz)

samplingrate: Sampling rate (Hz)

channelmodel: Channel model for path delays and average path gains: supporting TDL-C, and single Rayleigh path

rmsdelay: Root mean square delay for TDL channel models in nano seconds

maxorder: Maximum filter order for FDF

gainnormalization: Indicator for normalizing path gains

MultipathChannel.ChanPars — Type

MultipathChannel.ChanPars

Type of channel parameters containing:

numpaths::Int: Number of paths

pathdelays::Vector{Float64}: Delay of each path in second. This parameter is set according to channelmodel when calling MultipathChannel.initChan.

averagepathgains::Vector{Float64}: Average gain of each path in dB. This parameter is set according to channelmodel when calling MultipathChannel.initChan.

fdfcoefs::Vector{Vector{Float64}}: Coefficients of fractional delay filter (FDF). This parameter is obtained by pathdelays.

nonzeroindices::Vector{Int}: Non zero index of each path for FDF. This parameter is obtained by pathdelays.

pathdopplers::Vector{Float64}: Doppler frequency of each path in Hz. Doppler frequency is given according to Jakes model with maxdoppler. This parameter is initialized when calling initChan.

fadingamplitudes::Vector{Complex{Float64}}: Instantaneous gain of each path given by Rayleigh fading. This parameter is initialized when calling initChan.

initialphases::Vector{Float64}: Initial phase offset of each path. This parameter is initialized when calling initChan.

samplingrate::Float64: Sampling rate in Hz